Update app name in no access component
This fixes a reported bug where users lacking access to Activation Keys
were shown a message stating they don't have access to RHC. This was
due to using a shared component from RHC without updating the app name
to Activation Keys.

This also updates the Authentication test snapshot to reflect the new
application name so that tests pass as expected.