Check existing activation key names when creating a new one
This adds a check for existing activation key names to the create modal,
so that a user cannot attempt to create a key with a name that is
already used. It uses the activation keys hook, instead of passing a
prop to the modal because this is a federated module, and may need to
find its own information from somewhere else.

I've tested this filtering with 100,000 activation keys, and the
validation is still instant, so I am not worried about performance of
checking every key on each keystroke.