Use backend query parameters for frontend functions
This delegates all sorting, searching, and pagination to the backend
APIs that power the Add Additional Repositories modal. The query
parameters used are:

* limit (pageSize)
* offset (pageSize * page)
* repo_name
* repo_label
* sort_by
* sort_direction

All responses from unique pairings of these query parameters are cached,
so that as the user navigates this modal more, the quicker it becomes.

To enhance the user experience, the next page is always pre-fetched so
that if the user navigates in a normal use case (instead of just
spam-clicking the next page button) they should rarely see loaders as
they page through the modal.

This also debounces search inputs to only initiate a new query after
300ms of no-input. That way we do not constantly fire requests are
rhsmapi while the user types, and only search when they slow down.