Fix bug with toggling selected repositories
If repositories were selected, toggled to, and then anything cause a new
set of data to be loaded (for example, changing the search) we would
disregard the selection status of the repositories, allowing them to
display incorrectly, and be added a second (or third or twentieth time).

This puts two safeguards in place to prevent that.

1. Match selection on repository.repositoryLabel instead of a direct object
   reference

2. Ensure a repository with a matching label is not already selected
   before adding the current one to the selected repository list