Reintroduce extended update support product selection (#106)
* Reintroduct extended update support product selection

This fixes a bug where we accidentally removed the product selection
dropdown for extended update support activation keys. This reintroduces
it, and adds inference based on currently applied repos for _editing_
activation keys.

It also adds a helpful warning message in the case that we fail to infer
a product, telling the user what product is being defaulted to.

It also updates the warning message for editing the workload that
changing the version _or product_ may cause all currently applied
additional repos to be removed.

* Fix inference priority

* Fix inference bug

Only attempt to infer the product when a release version is already set
on the activation key.