TUSC-344: Generate UUID for activation key and warn user if changed (#159)
* Generate UUID for activation key and warn user if changed

This adds a default UUID name for an activation key so that they are
less guessable, in the case that the user customizes this name, it adds
a warning to let them know of potential risks.

This also revealed some flaws with the modal tests
(:old-man-yells-at-snapshot-testing:) so those have been updated as
well to actually test the pages that we were trying to.

* fix warning message typo